feat: Add three new validators for string length
This commit adds three new validators for matching string lengths - `isMinLength`, `isMaxLength` and
`isExactLength`. They all take a `length` parameter, which needs to be a number. They are great for
checking that a username is of a given length, that a zip code is valid, or that you don't have
people registering with a 200 character first name that breaks your database.